requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.15
django-filter==21.1
django-import-export==2.6.1
django-model-utils==4.0.0
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.35
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tablib==3.1.0
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
pandas==1.4.2